rename diagnostics -> evaluation; split scoring.py into evaluation/scoring.py and inference/scoring_configs.py
