+++ docs/api_reference/developer/evaluation/plotting_utils.md
@@ -2,17 +2,17 @@
 
 **plot_hmm_states_and_observations**
 
-::: dynestyx.diagnostics.plotting_utils.plot_hmm_states_and_observations
+::: dynestyx.evaluation.plotting_utils.plot_hmm_states_and_observations
     options:
       show_root_heading: false
       show_root_toc_entry: false
+      filters: []
 
 ---
 
 **plot_continuous_states_and_partial_observations**
 
-::: dynestyx.diagnostics.plotting_utils.plot_continuous_states_and_partial_observations
+::: dynestyx.evaluation.plotting_utils.plot_continuous_states_and_partial_observations
     options:
       show_root_heading: false
       show_root_toc_entry: false
-
