+++ docs/api_reference/developer/inference/scoring_configs.md
@@ -1,0 +1,10 @@
+# Scoring Configs
+
+`dynestyx.inference.scoring_configs` contains the `Filter`-specific configuration layer for predictive-observation scoring.
+
+The reusable score objects and policy types live on the companion [Scoring](../evaluation/scoring.md) page.
+
+::: dynestyx.inference.scoring_configs
+    options:
+      members:
+        - ObservationScoringConfig
