+++ docs/api_reference/public/inference/scoring_configs.md
@@ -1,0 +1,10 @@
+# Scoring Configs
+
+`ObservationScoringConfig` configures how `Filter(..., scoring_config=...)` applies scoring rules to predictive observation distributions.
+
+The score rule objects themselves live on the companion [Scoring](../evaluation/scoring.md) page.
+
+::: dynestyx.inference.scoring_configs
+    options:
+      members:
+        - ObservationScoringConfig
