--- dynestyx/diagnostics/__init__.py
@@ -1,1 +1,0 @@
-"""Diagnostics utilities for dynestyx."""
