+++ tests/test_science/test_hmm.py
@@ -8,7 +8,7 @@ import jax.random as jr
 import pytest
 from numpyro.infer import MCMC, NUTS
 
-from dynestyx.diagnostics.plotting_utils import plot_hmm_states_and_observations
+from dynestyx.evaluation.plotting_utils import plot_hmm_states_and_observations
 from tests.fixtures import data_conditioned_hmm  # noqa: F401
 from tests.test_utils import get_output_dir
